+++ src/main/java/insane96mcp/iguanatweaksreborn/network/MessageSaturationSync.java
@@ -1,0 +1,30 @@
+package insane96mcp.iguanatweaksreborn.network;
+
+import net.minecraft.network.PacketBuffer;
+import net.minecraftforge.fml.network.NetworkEvent;
+
+import java.util.function.Supplier;
+
+/*
+	Shamelessly copypasted from AppleSkin's code
+ */
+public class MessageSaturationSync {
+	float saturationLevel;
+
+	public MessageSaturationSync(float saturationLevel) {
+		this.saturationLevel = saturationLevel;
+	}
+
+	public static void encode(MessageSaturationSync pkt, PacketBuffer buf) {
+		buf.writeFloat(pkt.saturationLevel);
+	}
+
+	public static MessageSaturationSync decode(PacketBuffer buf) {
+		return new MessageSaturationSync(buf.readFloat());
+	}
+
+	public static void handle(final MessageSaturationSync message, Supplier<NetworkEvent.Context> ctx) {
+		ctx.get().enqueueWork(() -> NetworkHelper.getSidedPlayer(ctx.get()).getFoodStats().setFoodSaturationLevel(message.saturationLevel));
+		ctx.get().setPacketHandled(true);
+	}
+}